Move Open, New and NewTx to postgresql.go
